@@ -2,31 +2,29 @@
   "name": "robonaut-server-integration-test",
   "private": true,
   "dependencies": {
-    "axios": "1.2.2",
+    "axios": "1.2.4",
     "dotenv": "16.0.3",
     "pg": "8.8.0"
   },
   "devDependencies": {
     "@tsconfig/node16": "1.0.3",
-    "@types/jest": "29.2.5",
-    "@types/mkdirp": "1.0.2",
+    "@types/jest": "29.4.0",
     "@types/node": "ts3.7",
     "@types/pg": "8.6.6",
-    "@types/rimraf": "3.0.2",
-    "@typescript-eslint/eslint-plugin": "5.48.0",
-    "@typescript-eslint/parser": "5.48.0",
-    "eslint": "8.31.0",
+    "@typescript-eslint/eslint-plugin": "5.49.0",
+    "@typescript-eslint/parser": "5.49.0",
+    "eslint": "8.32.0",
     "eslint-config-prettier": "8.6.0",
     "eslint-plugin-jest": "27.2.1",
     "eslint-plugin-prettier": "4.2.1",
-    "jest": "29.3.1",
-    "jest-environment-node": "29.3.1",
+    "jest": "29.4.0",
+    "jest-environment-node": "29.4.0",
     "jest-junit": "15.0.0",
-    "mkdirp": "1.0.4",
-    "prettier": "2.8.2",
-    "rimraf": "3.0.2",
+    "mkdirp": "2.1.3",
+    "prettier": "2.8.3",
+    "rimraf": "4.1.2",
     "testcontainers": "9.1.1",
-    "ts-jest": "29.0.3",
+    "ts-jest": "29.0.5",
     "ts-node": "10.9.1",
     "typescript": "4.9.4"
   },
