@@ -1,9 +1,9 @@
-import rimraf from 'rimraf'
+import { rimraf } from 'rimraf'
 import jestTempDir from './jest-temp-dir'
 
 export default async function () {
   // remove temp folder
-  rimraf.sync(jestTempDir)
+  await rimraf(jestTempDir)
   // stop containers
   await globalThis.dockerComposeEnvironment.down()
 }
